fix(rust): clear the 56 clippy errors the new pipeline surfaced
Clippy had never run on this codebase, so `-D warnings` reported 56 errors on
first contact. Most are mechanical and fixed in place: eleven `max(0).min(1000)`
pagination guards become `clamp`, four `Err(e.into())` arms on `rusqlite::Error`
lose the no-op conversion, `MigrationStatus` derives `Default`, the union-find
grouping and the initial directory scan iterate instead of indexing, and the
`tract` model handle in `object_detection.rs` gets the same named type alias
`clip.rs` already uses.

The rest are deferred rather than fixed, because fixing them means changing
behaviour or deleting code that other tasks own. Each one carries an `#[allow]`
at the site with the reason and, where a task exists, its issue: `errors.rs` is
whole-module dead code until T32 (#40) wires typed errors up, `search_media`,
`get_persons` and `escape_like_pattern` are T55's (#63) to delete, and the two
`add_media*` signatures wait on the `database.rs` split in T58 (#66). The CLIP
normalisation constants keep their published precision, and `migrate` keeps its
final `version = N` so the next migration is appended rather than remembered.

`cargo clippy --all-targets -- -D warnings` and `cargo fmt --check` are clean.

+++ src-tauri/src/clip.rs
@@ -22,6 +22,9 @@ use tokenizers::Tokenizer;
 use tract_onnx::tract_hir::internal::*;
 
 /// CLIP embedding dimension (ViT-B/32)
+// Nothing validates embedding length against this today; kept as the documented shape of
+// what `clip_embedding` rows contain.
+#[allow(dead_code)]
 pub const EMBEDDING_DIM: usize = 512;
 
 /// Model filenames
@@ -337,13 +340,17 @@ pub fn encode_image(image_path: &Path) -> Result<Vec<f32>, String> {
 
     // Normalize and convert to NCHW
     // CLIP Mean and Std for normalization
+    // The published CLIP constants, kept as written upstream rather than truncated to the
+    // nearest `f32`, so they stay recognisable against the reference implementation.
+    #[allow(clippy::excessive_precision)]
     let mean = [0.48145466, 0.4578275, 0.40821073];
+    #[allow(clippy::excessive_precision)]
     let std = [0.26862954, 0.26130258, 0.27577711];
 
     let image_tensor: Tensor =
         tract_ndarray::Array4::from_shape_fn((1, 3, 224, 224), |(_, c, y, x)| {
             let pixel = resized.get_pixel(x as u32, y as u32);
-            let val = pixel[c as usize] as f32 / 255.0;
+            let val = pixel[c] as f32 / 255.0;
             (val - mean[c]) / std[c]
         })
         .into();
